+++ src/adaptors/index.ts
@@ -10,6 +10,7 @@ export { default as getContextData } from './context';
 export { default as getCoordinapeData } from './coordinape';
 export { default as getCryptoscamdbData } from './cryptoscamdb';
 export { default as getCyberconnectData } from './cyberconnect';
+export { default as getDapplistData } from './dapplist';
 export { default as getDeepDaoData } from './deepdao';
 export { default as addressToEns } from './ens';
 export { default as getEtherscanData } from './etherscan';